Split `handle-page-function.js` and `extraction.js` into separate files
